move watcher and setter to stepper (#213)
